fix: 階層の移動とHPのテキストを表示
issue: 148
Notes:
- HPバーの一式をCharacterの階層へ移動
- 同時に名前空間もUraty.Features.Characterになりました
- ダメージと回復のテキスト描画未実装
