+++ Assets/_Features/Character/Prefabs/BarBase.prefab.meta
