+++ Assets/_Features/Character/Prefabs/HPBar.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 2c53cf9a2e1a3a94aab4365816ac7ae1
-folderAsset: yes
-DefaultImporter:
+guid: 116e9a2fd785230468c4046cfe3199ae
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
