+++ Assets/_Features/Character/Scripts/BarBase.cs.meta
