+++ Assets/_Features/Character/UI/HPBar.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ccbe8589480806046be4fee8965e1e9c
+guid: 008dc6812611abf4b97976801a4770b2
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
