--- Assets/_Features/Character/UI/HPBar/HPBarDamegeReceive.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a2793056fca194368d711a9565611f28a099aef00b3aa873971514ac8a5c4df3
-size 54012
