+++ Assets/_Features/Character/UI/HPBar/Textures/HPBarDamegeReceive.png.meta
