+++ Assets/_Features/Character/UI/HPBar/Textures/HPBarMatcha.png.meta
