+++ Assets/_Features/Character/UI/HPBar/Textures/HPBarStrawberry.png.meta
