+++ Assets/_Features/Character/UI/HPBar/Textures/HPBarStrawberryLow.png.meta
