--- Assets/_Sandboxes/Nishiguchi/HPBar.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 008dc6812611abf4b97976801a4770b2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
