--- Assets/_Sandboxes/Nishiguchi/HPBar/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5f92c47c2b5978a4f9728e333e126043
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
