--- Assets/_Sandboxes/Nishiguchi/HPBar/Textures/HPBarBackGround.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:cdd45bd8519b92c47ff731ccb1b0414b5b37a307e260cfd712184a4f299da177
-size 5791
