--- Assets/_Sandboxes/Nishiguchi/HPBar/Textures/HPBarDamageCause.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c00a13d75bb1e7eaaf84b1a1175a2fc822b3b254b8ae944a93c16d0ec5e41161
-size 39402
