--- Assets/_Sandboxes/Nishiguchi/HPBar/Textures/HPBarMatcha.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:36b5ff5ac7e74eff2a5aa2b5ff387ffb5c46fb4af55b656b6f31fa742ac0508d
-size 24768
