--- Assets/_Sandboxes/Nishiguchi/HPBar/Textures/HPBarStrawberry.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d7b79c51eaa6821f54e8442a2359a72cf1c6bc2d30887083f563d723988926f0
-size 25541
