--- Assets/_Sandboxes/Nishiguchi/HPBar/Textures/HPBarStrawberryLow.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c321d69b9fbbef1e971c9c155b0301339afb948b6cfb903f3bf4a0578da5fb47
-size 24887
