--- Assets/_Sandboxes/Nishiguchi/HPBar/Uraty.Feature.Health.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "Uraty.Feature.Health",
-    "rootNamespace": "Uraty.Feature.Health",
-    "references": [
-        "GUID:75469ad4d38634e559750d17036d5f7c"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
