--- Assets/_Sandboxes/Nishiguchi/HPBar/Uraty.Feature.Health.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c10a527a8863e4245b075e75570c26ef
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
